Move file structure to handle example files
