+++ components/gly_gen/api/glycosight/data/01CPTAC_OVprospective_G_JHUZ_20160317_QE_r01.mzid.gz
