+++ components/gly_gen/api/glycosight/index.ts
@@ -6,6 +6,7 @@ export const exampleFile = path.resolve(
   process.env.APP_ROOT as string,
   'components',
   'gly_gen',
+  'api',
   'glycosight',
   'data',
   '01CPTAC_OVprospective_G_JHUZ_20160317_QE_r01.mzid.gz',
@@ -14,9 +15,8 @@ export const exampleFile = path.resolve(
 export type UploadExampleFileResponse = Awaited<ReturnType<typeof uploadExampleFile>>
 
 export async function uploadExampleFile(session?: SessionWithId) {
-  console.log("=== I've been called! ===")
   return await uploadFile({
-    url: `${process.env.PUBLIC_URL}/api/v1/components/data/gly_gen/glycosight/data`,
+    url: `${process.env.PUBLIC_URL}/api/v1/components/gly_gen/glycosight`,
     filename: '01CPTAC_OVprospective_G_JHUZ_20160317_QE_r01.mzid.gz',
   }, session)
 }
